refactor : user 탈퇴, api code 수정
